Add files via upload
Second Paper Implementation. The following program implements the paper released by Meta discussing the Segment Everything Model. Here, I focus on a couple ways to increase efficiency of mask by annotating images. Mainly inspired by previous 1-on-1 as it was mentioned that it is likely that we will have to annotate dataset prior to model implementation.